normalize attacker countries data in useWidgetData and update charts
